refactor: update import paths for utility functions to improve organization and maintainability across various components and sections
