+++ app/routes/($locale).blogs.$blogHandle._index.tsx
@@ -4,7 +4,7 @@ import {json} from '@shopify/remix-oxygen';
 import {type RouteLoaderArgs} from '@weaverse/hydrogen';
 import {routeHeaders} from '~/data/cache';
 import {BLOGS_PAGE_QUERY} from '~/data/queries';
-import {PAGINATION_SIZE} from '~/lib/const';
+import {PAGINATION_SIZE} from '~/lib/utils/const';
 import {seoPayload} from '~/lib/seo.server';
 import {WeaverseContent} from '~/weaverse';
 import type {BlogQuery} from 'storefrontapi.generated';
