+++ app/routes/($locale).products.$handle.tsx
@@ -21,7 +21,7 @@ import {
 } from '~/data/queries';
 import {createJudgemeReview, getJudgemeReviews} from '~/lib/judgeme';
 import {seoPayload} from '~/lib/seo.server';
-import type {Storefront} from '~/lib/type';
+import type {Storefront} from '~/lib/types/type-locale';
 import {WeaverseContent} from '~/weaverse';
 import {useEffect} from 'react';
 import type {ProductRecommendationsQuery} from 'storefrontapi.generated';
