+++ app/routes/($locale).products._index.tsx
@@ -8,7 +8,7 @@ import type {LoaderFunctionArgs} from '@shopify/remix-oxygen';
 import {json} from '@shopify/remix-oxygen';
 import {routeHeaders} from '~/data/cache';
 import {ALL_PRODUCTS_QUERY} from '~/data/queries';
-import {PAGINATION_SIZE} from '~/lib/const';
+import {PAGINATION_SIZE} from '~/lib/utils/const';
 import {seoPayload} from '~/lib/seo.server';
 import {WeaverseContent} from '~/weaverse';
 import invariant from 'tiny-invariant';
